Rename addtional_transformers -> cbpm
